+++ src/aui-tooltip/tests/unit/js/aui-tooltip-delegate-tests.js
@@ -13,9 +13,11 @@ YUI.add('aui-tooltip-delegate-tests', function(Y) {
 
         'should show tooltips for all triggers': function() {
             var tooltip,
+                tooltipClass = 'tooltip-class',
                 triggers = Y.all('.tooltip-btn');
 
             this._delegate = new Y.TooltipDelegate({
+                cssClass: tooltipClass,
                 trigger: '.tooltip-btn'
             });
 
@@ -29,14 +31,27 @@ YUI.add('aui-tooltip-delegate-tests', function(Y) {
                     tooltip.get('trigger'),
                     'Trigger should have been updated'
                 );
+
                 Y.Assert.isFalse(
                     tooltip.get('boundingBox').hasClass('tooltip-hidden'),
                     'Tooltip should not be hidden after mouseover'
                 );
+
                 Y.Assert.areEqual(
                     trigger.getAttribute('data-title'),
                     tooltip.get('contentBox').get('text'),
                     'Tooltip content should have been updated for this trigger'
+                );
+
+                Y.Assert.isTrue(
+                    tooltip.get('boundingBox').hasClass(tooltipClass),
+                    'Tooltip should have the class assigned in the constructor'
+                );
+
+                Y.Assert.areEqual(
+                    tooltipClass,
+                    tooltip.get('cssClass'),
+                    'Tooltip object should have the assigned \'cssClass\' attribute'
                 );
 
                 trigger.simulate('mouseout');