+++ src/aui-datepicker/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-1205](https://issues.liferay.com/browse/AUI-1205) datePicker fires selectionChange event when trigger is clicked
 
 ## [3.0.0](https://github.com/liferay/alloy-ui/releases/tag/3.0.0)