+++ src/aui-scheduler/HISTORY.md
@@ -4,6 +4,8 @@
 
 ## @VERSION@
 
+* [AUI-1640](https://issues.liferay.com/browse/AUI-1640) Scheduler should create SchedulerEvent instances lazily
+* [AUI-1771](https://issues.liferay.com/browse/AUI-1771) Display all events in scheduler's month view
 * [AUI-1772](https://issues.liferay.com/browse/AUI-1772) Improve performance of SchedulerTableView plotEvents method
 
 ## [3.0.0](https://github.com/liferay/alloy-ui/releases/tag/3.0.0)